feat: impl caching
- minor refactoring
